Fix exception messages to use FieldElement.getReal
Previously severl exceptions used FieldElement.toString(), which is not
helpful for many types of Field. It just gives the objec type and
hashcode. Scrubbed uses of OrekitException to find places where a
FieldElement was formatted into the message. Updated those to use
getReal() or toAbsoluteDate() to make a more helpful message.